% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_leaf_area.R
\name{map_leaf_area}
\alias{map_leaf_area}
\title{Leaf area calculation loop}
\usage{
map_leaf_area(
  files,
  temp.folder = NULL,
  output.folder = NULL,
  imagej.path = NULL,
  pixel = 300,
  know.pixel.distance = 2.54,
  low.size = 0.005,
  trim.pixel = 100,
  trim.pixel2 = 150,
  ...
)
}
\arguments{
\item{files}{A character vector of file paths}

\item{temp.folder}{The folder where the files should be copied to}

\item{output.folder}{The folder where the output files should be saved}

\item{imagej.path}{The path to the ImageJ program. On Windows this will always be "C:\\Program Files\\ImageJ\\".}

\item{pixel}{The number of pixels in a known distance}

\item{know.pixel.distance}{The known distance in cm}

\item{low.size}{The minimum size of a leaf in cm2}

\item{trim.pixel}{The number of pixels to trim}

\item{trim.pixel2}{The number of pixels to trim}

\item{...}{aditional argument of run_ij.}
}
\value{
A dataframe containing the ID and LeafArea
}
\description{
Loop through a list of files and run ImageJ to measure leaf area.
}
\examples{
map_leaf_area(files = c("file1.tif", "file2.tif"),
temp.folder = NULL, output.folder = NULL,
imagej.path = NULL,
pixel = 300, know.pixel.distance = 2.54, low.size = 0.005,
trim.pixel = 100, trim.pixel2 = 150)
}
