% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_img_scans.R
\name{read_img_scans}
\alias{read_img_scans}
\title{Read image scans from a folder}
\usage{
read_img_scans(img_folder)
}
\arguments{
\item{img_folder}{A character string giving the path to the root folder containing the image scans.}
}
\value{
Returns a vector containing the normalized path to all image scans.
}
\description{
Read image scans from a folder
}
\examples{
read_img_scans("/home/user/data/")
}
