% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resmerge_ij.R
\name{resmerge.ij}
\alias{resmerge.ij}
\title{File management}
\usage{
resmerge.ij(path, prefix = "\\\\.|-")
}
\arguments{
\item{path}{Path to the target directory}

\item{prefix}{Regular expression to manage file names}
}
\value{
A data frame of total leaf area for each sample

\item{sample}{Name of sample}

\item{total.leaf.area}{Total leaf area of the sample (cm2)}
}
\description{
File management function. The output file contains sample names in the first column and total leaf area (cm2) of the sample (e.g., one individual plant or one ramet) in the second column.
}
\author{
Masatoshi Katabuchi \email{mattocci27@gmail.com}
}
