% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_leaf_area.R
\name{summarise_leaf_area}
\alias{summarise_leaf_area}
\title{Summarise Leaf Area}
\usage{
summarise_leaf_area(df)
}
\arguments{
\item{df}{A dataframe containing two columns (scan_id and leaf_area_cm2)}
}
\value{
A dataframe containing two columns (id and leaf_area_cm2)
}
\description{
Summarise Leaf Area
}
\examples{
# summarise_leaf_area(df)
}
